+++ mss/check-intra-redzone.cpp
@@ -18,10 +18,10 @@ int main(int argc, char* argv[])
   long long length = 0;
 
   switch(store_type) {
-    case 0: GET_DISTANCE(length, buffer_stack.data, buffer_stack.underflow); break;
-    case 1: GET_DISTANCE(length, buffer_heap->data, buffer_heap->underflow); break;
-    case 2: GET_DISTANCE(length, buffer_data.data, buffer_data.underflow); break;
-    case 3: GET_DISTANCE(length, buffer_rodata.data, buffer_rodata.underflow); break;
+    case 0: GET_DISTANCE(length, (long long )buffer_stack.data, (long long )buffer_stack.underflow); break;
+    case 1: GET_DISTANCE(length, (long long )buffer_heap->data, (long long )buffer_heap->underflow); break;
+    case 2: GET_DISTANCE(length, (long long )buffer_data.data, (long long )buffer_data.underflow); break;
+    case 3: GET_DISTANCE(length, (long long )buffer_rodata.data, (long long )buffer_rodata.underflow); break;
   }
   delete buffer_heap;
   long long abs_len = llabs(length);
