+++ mss/read-cross-frame-by-index.cpp
@@ -6,7 +6,7 @@
 int FORCE_NOINLINE helper(const char* b) {
   long long delta = 0;
   charBuffer dummy;
-  GET_DISTANCE(delta, b, dummy.data);
+  GET_DISTANCE(delta, (long long)b, (long long)dummy.data);
   return read_by_index(dummy, delta, 7, 1, 'd');
 }
